app/requirements.txt
uvicorn
fastapi
python-dateutil
pytest
autopep8
python-multipart
flake8
chardet
requests
bs4